test: revise tests
