+++ strands-ts/src/agent/__tests__/agent-as-tool.test.ts
@@ -5,7 +5,7 @@ import { MockMessageModel } from '../../__fixtures__/mock-message-model.js'
 import { collectGenerator } from '../../__fixtures__/model-test-helpers.js'
 import { createMockContext } from '../../__fixtures__/tool-helpers.js'
 import { ToolValidationError } from '../../errors.js'
-import { Tool, ToolStreamEvent } from '../../tools/tool.js'
+import { Tool, ToolStreamEvent, DIRECT_RETURN_DESCRIPTION_SUFFIX } from '../../tools/tool.js'
 import { ToolResultBlock } from '../../types/messages.js'
 import { SessionManager } from '../../session/session-manager.js'
 import type { SnapshotStorage } from '../../session/storage.js'
@@ -405,6 +405,28 @@ describe('AgentAsTool', () => {
 
       expect(tool.name).toBe('custom-name')
       expect(tool.description).toBe('Custom desc')
+    })
+
+    it('sets directReturn and appends description suffix when handoff is true', () => {
+      const model = new MockMessageModel().addTurn({ type: 'textBlock', text: 'Hi' })
+      const agent = new Agent({ model, name: 'specialist', description: 'Handles billing', printer: false })
+
+      const tool = agent.asTool({ handoff: true })
+
+      expect(tool.directReturn).toBe(true)
+      expect(tool.description).toBe('Handles billing' + DIRECT_RETURN_DESCRIPTION_SUFFIX)
+      expect(tool.toolSpec.description).toBe('Handles billing' + DIRECT_RETURN_DESCRIPTION_SUFFIX)
+    })
+
+    it('does not set directReturn or append description suffix when handoff is false', () => {
+      const model = new MockMessageModel().addTurn({ type: 'textBlock', text: 'Hi' })
+      const agent = new Agent({ model, name: 'specialist', description: 'Handles billing', printer: false })
+
+      const tool = agent.asTool({ handoff: false })
+
+      expect(tool.directReturn).toBe(false)
+      expect(tool.description).toBe('Handles billing')
+      expect(tool.description).not.toContain(DIRECT_RETURN_DESCRIPTION_SUFFIX)
     })
   })
